5.5.0
Changes:

* Removed PROGUARD Obfuscation because the project heavily relies on
Reflection and Enumeration which will be scrambled when obfuscated

* Removed Keystore from Commits

* Changed Version name shown in Settings > Apps > SharpFix Android >
Version

* Target SDK upgraded to API 19 (Kitkat)

* Added topMostPath to FileDialog to avoid navigating unrelated
directories such as root directories (even if the device is rooted)
except in developer mode

* Revised the way ApplicationClass is doing its Active and Mounted
Volumes Detection. This feature can now be called globally across the
application

* Changed {DIR} + {FILE} format in FileDialog to (ImageView + TextView
format), avoiding confusion for Files and Directories/Folders

* Fixed the bug in Logcat.java that it throws exception whenever the
argument stacktrace is null

* Added Cancel button in FileDialog, to close the Dialog.

* Changed FileDialog navigate parent function from clicking the ".." in
the list to simply pressing the back button.